pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pr-reviewer"
version = "0.1.0"
description = "Agentic PR review tool powered by Claude — rivals CodeRabbit, Qodo, Greptile"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.40",
    "PyGithub>=2.3",
    "atlassian-python-api>=3.41",
    "requests>=2.32",
    "pydantic>=2.9",
    "typer>=0.13",
    "rich>=13.9",
    "tree-sitter>=0.23",
    "tree-sitter-language-pack>=0.1",
    "networkx>=3.4",
    "diskcache>=5.6",
    "unidiff>=0.7",
    "pyyaml>=6.0",
    "structlog>=24.4",
    "ratelimit>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-recording>=0.13",
    "vcrpy>=6.0",
    "pytest-mock>=3.14",
    "httpx>=0.27",
    "mypy>=1.11",
    "ruff>=0.6",
]

[project.scripts]
pr-reviewer = "pr_reviewer.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pr_reviewer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "unit: pure unit tests, no network",
    "integration: VCR cassette replay",
    "e2e: live network tests, gated by RUN_E2E=1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
